+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.11.0] - 2023-12-30
 ### Added
  - Show app icons for custom shortcuts in the app selector
  - Verbose flag can be enabled with `-vv` for additional debug logging
@@ -278,7 +278,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.10.5...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.11.0...HEAD
+[1.11.0]: https://github.com/Matoking/protontricks/compare/1.10.5...1.11.0
 [1.10.5]: https://github.com/Matoking/protontricks/compare/1.10.4...1.10.5
 [1.10.4]: https://github.com/Matoking/protontricks/compare/1.10.3...1.10.4
 [1.10.3]: https://github.com/Matoking/protontricks/compare/1.10.2...1.10.3